Filling dictionary from serialised API
